+++ packages/x-charts/src/internals/index.ts
@@ -30,6 +30,7 @@ export * from '../context/ColorProvider';
 export * from '../context/InteractionProvider';
 export * from '../context/SeriesProvider';
 export * from '../context/ZAxisContextProvider';
+export * from '../context/PluginProvider';
 export type * from '../context/context.types';
 
 // series configuration
