+++ packages/x-charts/src/models/index.ts
@@ -1,7 +1,6 @@
 export * from './seriesType';
 export * from './layout';
 export * from './stacking';
-export * from './plugin';
 export type {
   AxisConfig,
   ChartsYAxisProps,
